Fixed broken signal warning on Qt 4

+++ src/opl/generator.cpp
@@ -709,9 +709,9 @@ void Generator::changePatch(FmBank::Instrument &instrument, bool isDrum)
     }
 }
 
-void Generator::changeNote(int32_t newnote)
+void Generator::changeNote(int newnote)
 {
-    note = newnote;
+    note = int32_t(newnote);
 }
 
 void Generator::changeDeepTremolo(bool enabled)
